backend/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
cohere>=4.0.0
qdrant-client>=1.8.0
python-dotenv>=1.0.0
Flask>=2.3.0
google-generativeai>=0.4.0
openai>=1.0.0